requirements.txt

requests==2.31.0
selenium==4.17.2
schedule==1.2.1
webdriver-manager==4.0.1
python-dotenv==1.0.0